CT-48 Auth block is transfered in comnponent
